requirements.txt
# HEAL - Requirements

# Face detection and landmarks
mediapipe>=0.10.0

# Computer vision
opencv-python>=4.8.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Machine learning for classification (optional)
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep learning for CNN-based classification (optional)
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
